no inheritance at all and move service to orders app
