--- src/products/services/__init__.py
@@ -1,3 +1,0 @@
-__all__ = ["PurchaseCreator"]
-
-from products.services.purchase_creator import PurchaseCreator
