+++ src/products/urls.py
@@ -1,9 +1,9 @@
 from django.urls import path
 
-from products.api.views import CoursePromocodeView
-from products.api.views import CoursePurchaseView
+from products.api.views import PromocodeView
+from products.api.views import PurchaseView
 
 urlpatterns = [
-    path("courses/<str:slug>/promocode/", CoursePromocodeView.as_view()),
-    path("courses/<str:slug>/purchase/", CoursePurchaseView.as_view()),
+    path("courses/<str:slug>/promocode/", PromocodeView.as_view()),
+    path("courses/<str:slug>/purchase/", PurchaseView.as_view()),
 ]
